requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
click==8.0.3
dj-rest-auth==2.2.1
Django==4.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
